refactor(core): unify mention/slash under Trigger + Directive; harden parser
Consolidate the Mention and SlashCommand type aliases into a single Trigger + Directive model. Rename mention-formatter to directive-formatter. Bound the directive regex (type ≤64, label/name ≤1024) and forbid newlines so pathological inputs can't make parsing cross line boundaries.
