+++ packages/core/src/index.ts
@@ -42,11 +42,9 @@ export type { Unsubscribe } from "./types/unsubscribe";
 export type { QuoteInfo } from "./types/quote";
 
 export type {
-  Unstable_MentionItem,
-  Unstable_MentionCategory,
   Unstable_DirectiveSegment,
   Unstable_DirectiveFormatter,
-} from "./types/mention";
+} from "./types/directive";
 
 export type {
   Unstable_TriggerItem,
@@ -120,15 +118,8 @@ export type { FeedbackAdapter } from "./adapters/feedback";
 // Suggestion adapter
 export type { SuggestionAdapter } from "./adapters/suggestion";
 
-// Mention adapter
-export type { Unstable_MentionAdapter } from "./adapters/mention";
-export { unstable_defaultDirectiveFormatter } from "./adapters/mention";
-
-// Slash command adapter
-export type {
-  Unstable_SlashCommandAdapter,
-  Unstable_SlashCommandItem,
-} from "./adapters/trigger";
+// Directive formatter
+export { unstable_defaultDirectiveFormatter } from "./adapters/directive-formatter";
 
 // Thread history adapters
 export type {
